+++ test/TestHelper/TestHelper.psm1
@@ -34,7 +34,7 @@ function Connect-AzureServicePrincipal
         $ApplicationId,
 
         [Parameter(Mandatory = $true)]
-        [System.String]
+        [System.Security.SecureString]
         $ApplicationPassword,
 
         [Parameter(Mandatory = $true)]
@@ -47,13 +47,9 @@ function Connect-AzureServicePrincipal
         Write-Verbose -Message "Logging in to Azure using Service Principal $ApplicationId"
 
         # Build platform (AppVeyor) does not offer solution for passing secure strings
-        $secureStringPassword = ConvertTo-SecureString `
-            -String $ApplicationPassword `
-            -AsPlainText `
-            -Force
         $azureCredential = New-Object `
             -Typename System.Management.Automation.PSCredential `
-            -Argumentlist $ApplicationId, $secureStringPassword
+            -Argumentlist $ApplicationId, $applicationPassword
 
         # Suppress request to share usage information
         $path = "$Home\AppData\Roaming\Windows Azure Powershell\"
@@ -64,12 +60,12 @@ function Connect-AzureServicePrincipal
         $azureProfileFilename = Join-Path `
             -Path $Path `
             -ChildPath 'AzureDataCollectionProfile.json'
-        $azureProfileContent = Set-Content `
+        $null = Set-Content `
             -Value '{"enableAzureDataCollection":true}' `
             -Path $azureProfileFilename
 
         # Handle login
-        $null = Add-AzAccount `
+        $null = Add-AzureRmAccount `
             -ServicePrincipal `
             -SubscriptionId $SubscriptionId `
             -TenantId $TenantId `
@@ -77,7 +73,7 @@ function Connect-AzureServicePrincipal
             -ErrorAction SilentlyContinue
 
         # Validate login
-        $loginSuccessful = Get-AzSubscription `
+        $loginSuccessful = Get-AzureRmSubscription `
             -SubscriptionId $SubscriptionId `
             -TenantId $TenantId
 
@@ -123,7 +119,7 @@ function New-AzureTestCosmosDbAccount
         }
 
         # Deploy ARM template
-        New-AzResourceGroupDeployment `
+        New-AzureRmResourceGroupDeployment `
             @deploymentParameters
     }
     catch [System.Exception]
@@ -151,7 +147,7 @@ function Remove-AzureTestCosmosDbAccount
         Write-Verbose -Message ('Removing Cosmos DB test account {0}.' -f $AccountName)
 
         # Remove resource group as
-        $null = Remove-AzResourceGroup `
+        $null = Remove-AzureRmResourceGroup `
             -Name $ResourceGroupName `
             -Force `
             -AsJob
@@ -180,7 +176,7 @@ function New-AzureTestCosmosDbResourceGroup
     {
         Write-Verbose -Message ('Creating test Azure Resource Group {0} in {1}.' -f $ResourceGroupName,$Location)
 
-        $null = New-AzResourceGroup `
+        $null = New-AzureRmResourceGroup `
             -Name $ResourceGroupName `
             -Location $Location
     }
@@ -204,7 +200,7 @@ function Remove-AzureTestCosmosDbResourceGroup
     {
         Write-Verbose -Message ('Removing test Azure Resource Group {0}.' -f $ResourceGroupName)
 
-        $null = Remove-AzResourceGroup `
+        $null = Remove-AzureRmResourceGroup `
             -Name $ResourceGroupName `
             -Force `
             -AsJob
@@ -221,7 +217,7 @@ function Convert-SecureStringToString
     param
     (
         [Parameter(Mandatory = $true, ValueFromPipeline = $true)]
-        [SecureString]
+        [System.Security.SecureString]
         $SecureString
     )